requirements.txt
flask
flask-mysql
flask-cors
flask-mysql-connector
